Moved config to internal
